Crate deskulpt_build

This crate implements utilities for managing the repository workspace of Deskulpt.

⚠️ This crate is meant to be used internally by the Deskulpt application and is not designed to support plugin authors or other users directly. Private items are documented for reference of Deskulpt developers.

Structs

Builder

Builder for build-time configuration of Deskulpt.